Modify ping wait time
The previous command sets the ping wait time
to 500 seconds, so when the jetson is disconnected
it takes way too long for the script to show
NOT Connected, as it has to wait until ping
exits. Now, when the jetson is disconnected it should
(in theory) take 4 seconds between NOT Connected messages
since it's 2 seconds sleep then 2 seconds ping.